+++ Common/src/main/java/fuzs/netherchested/world/level/block/NetherChestBlock.java
@@ -14,7 +14,6 @@ import net.minecraft.network.chat.Component;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.util.RandomSource;
-import net.minecraft.world.Container;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResult;
 import net.minecraft.world.MenuProvider;
@@ -31,7 +30,6 @@ import net.minecraft.world.level.block.EnderChestBlock;
 import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.entity.BlockEntityTicker;
 import net.minecraft.world.level.block.entity.BlockEntityType;
-import net.minecraft.world.level.block.entity.ChestBlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.phys.BlockHitResult;
 
@@ -41,7 +39,8 @@ import java.util.OptionalInt;
 
 @SuppressWarnings("deprecation")
 public class NetherChestBlock extends EnderChestBlock {
-    private static final Component DESCRIPTION_COMPONENT = Component.translatable("block.netherchest.nether_chest.description").withStyle(ChatFormatting.GOLD);
+    public static final String NETHER_CHEST_DESCRIPTION_KEY = "block.netherchested.nether_chest.description";
+    private static final Component DESCRIPTION_COMPONENT = Component.translatable(NETHER_CHEST_DESCRIPTION_KEY).withStyle(ChatFormatting.GOLD);
 
     public NetherChestBlock(Properties properties) {
         super(properties);
@@ -50,9 +49,8 @@ public class NetherChestBlock extends EnderChestBlock {
     @Override
     public void setPlacedBy(Level level, BlockPos pos, BlockState state, LivingEntity placer, ItemStack stack) {
         if (stack.hasCustomHoverName()) {
-            BlockEntity blockEntity = level.getBlockEntity(pos);
-            if (blockEntity instanceof ChestBlockEntity) {
-                ((ChestBlockEntity) blockEntity).setCustomName(stack.getHoverName());
+            if (level.getBlockEntity(pos) instanceof NetherChestBlockEntity blockEntity) {
+                blockEntity.setCustomName(stack.getHoverName());
             }
         }
     }
@@ -60,9 +58,8 @@ public class NetherChestBlock extends EnderChestBlock {
     @Override
     public void onRemove(BlockState state, Level level, BlockPos pos, BlockState newState, boolean isMoving) {
         if (!state.is(newState.getBlock())) {
-            BlockEntity blockEntity = level.getBlockEntity(pos);
-            if (blockEntity instanceof Container) {
-                UnlimitedContainerUtils.dropContents(level, pos, (Container) blockEntity);
+            if (level.getBlockEntity(pos) instanceof NetherChestBlockEntity blockEntity) {
+                UnlimitedContainerUtils.dropContents(level, pos, blockEntity.container);
                 level.updateNeighbourForOutputSignal(pos, this);
             }
 
@@ -93,8 +90,8 @@ public class NetherChestBlock extends EnderChestBlock {
     }
 
     @Override
-    public BlockEntity newBlockEntity(BlockPos pPos, BlockState pState) {
-        return new NetherChestBlockEntity(pPos, pState);
+    public BlockEntity newBlockEntity(BlockPos pos, BlockState state) {
+        return ModRegistry.NETHER_CHEST_BLOCK_ENTITY_TYPE.get().create(pos, state);
     }
 
     @Override
@@ -110,14 +107,16 @@ public class NetherChestBlock extends EnderChestBlock {
 
     @Override
     public int getAnalogOutputSignal(BlockState state, Level level, BlockPos pos) {
-        return UnlimitedContainerUtils.getRedstoneSignalFromBlockEntity(level.getBlockEntity(pos));
+        if (level.getBlockEntity(pos) instanceof NetherChestBlockEntity blockEntity) {
+            return UnlimitedContainerUtils.getRedstoneSignalFromContainer(blockEntity.container);
+        }
+        return 0;
     }
 
     @Override
     public void tick(BlockState state, ServerLevel level, BlockPos pos, RandomSource random) {
-        BlockEntity blockEntity = level.getBlockEntity(pos);
-        if (blockEntity instanceof ChestBlockEntity) {
-            ((ChestBlockEntity) blockEntity).recheckOpen();
+        if (level.getBlockEntity(pos) instanceof NetherChestBlockEntity blockEntity) {
+            blockEntity.recheckOpen();
         }
     }