+++ Forge/src/main/java/fuzs/netherchested/data/ModLanguageProvider.java
@@ -1,6 +1,7 @@
 package fuzs.netherchested.data;
 
 import fuzs.netherchested.init.ModRegistry;
+import fuzs.netherchested.world.level.block.NetherChestBlock;
 import net.minecraft.data.DataGenerator;
 import net.minecraftforge.common.data.LanguageProvider;
 
@@ -13,7 +14,7 @@ public class ModLanguageProvider extends LanguageProvider {
     @Override
     protected void addTranslations() {
         this.add(ModRegistry.NETHER_CHEST_BLOCK.get(), "Nether Chest");
+        this.add(NetherChestBlock.NETHER_CHEST_DESCRIPTION_KEY, "Allows for storing way more items in a single slot than any other inventory.");
         this.add("container.nether_chest", "Nether Chest");
-        this.add("block.netherchest.nether_chest.description", "Allows for storing way more items in a single slot than any other inventory.");
     }
 }